## 변경 내용 serializers2.py -> serializers.py views2.py , views3.py -> views.py 사용자 계정 조회 , 수정 , 삭제 일때 로그인 여부 , request user = token user 확인 여부 , body/url/get 으로 얻어오는 데이터 정보와 request user 가 맞는지 여부 기능 추가 model 수정 choice field 수정

+++ server/api/views.py
@@ -8,9 +8,8 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
-from .models import User
-from .serializers import UserDetailSerializer, UserCreateSerializer, UserUpdateSerializer, UserFindIDSerializer, \
-    UserPasswordSerializer, UserChangePasswordSerializer
+
+from .serializers import *
 from .permissions import IsTokenOwner
 
 
@@ -24,6 +23,8 @@ class UserView(APIView):
     @action(detail=True, methods=['get'], permission_classes=[IsAuthenticated, IsTokenOwner])  # get 함수를 사용할때 적용하는 권한 설정
     def get(self, request):
         id = request.query_params.get('id')  # url 에서 받아온 id 값
+        if id != request.user.id:
+            return Response(data={"state":"다른 이용자의 정보입니다 열람 하실수 없습니다."})
         try:
             user = User.objects.get(id=id)  # 정보를 보고 싶은 사용자 레코드
             serializer = UserDetailSerializer(user)  # 직렬화
@@ -43,12 +44,14 @@ class UserView(APIView):
     # 회원정보 업데이트
     @action(detail=True, methods=['put'], permission_classes=[IsAuthenticated, IsTokenOwner])
     def put(self, request):
+        id = request.data.get('id') # body 에 존재하는 이용자가 요청한 id 값
+
+        if id != request.user.id:   # id 와 request 요청으로 들어온 id 의 정보가 다르면 수정 불가
+            return Response(data={"state": "다른 사용자의 아이디 입니다 수정할수 없습니다"}, status=status.HTTP_400_BAD_REQUEST)
         try:
-            id = request.data.get('id')
-            if id != request.user:
-                return Response(data={"state":"다른 사용자의 아이디 입니다 수정할수 없습니다"} , status=status.HTTP_400_BAD_REQUEST)
             user = User.objects.get(id=id)  # 업데이트 할 사용자 정보 조회
-            serializer = UserUpdateSerializer(user, data=request.data)  # 해당 사용자 정보 직렬화 + 유저가 변경하고 싶은 정보
+            serializer = UserUpdateSerializer(user, data=request.data,
+                                              context={'request': request})  # 해당 사용자 정보 직렬화 + 유저가 변경하고 싶은 정보
 
             if serializer.is_valid():  # 유효성 검사
                 serializer.update(user, serializer.validated_data)  # 사용자 정보 업데이트
@@ -59,8 +62,11 @@ class UserView(APIView):
             return Response(data={"state": "회원 정보가 존재하지 않습니다."}, status=status.HTTP_404_NOT_FOUND)
 
     # 회원 정보 delete
-    @action(detail=True, methods=['delete'], permission_classes=[IsAuthenticated,IsTokenOwner])  # get 함수를 사용할때 적용하는 권한 설정
+    @action(detail=True, methods=['delete'],
+            permission_classes=[IsAuthenticated, IsTokenOwner])  # get 함수를 사용할때 적용하는 권한 설정
     def delete(self, request, id):
+        if id != request.user.id:   # url 에서 받아온 id 와 request 요청으로 들어온 아이디가 다를시에도 탈퇴 불가
+            return Response(data={"state": "다른 사용자의 아이디 입니다. 계정 탈퇴할수 없습니다"}, status=status.HTTP_400_BAD_REQUEST)
         try:
             user = User.objects.get(id=id)  # 삭제하고 싶은 사용자 정보 조회
             user.delete()  # 사용자 정보 삭제
@@ -220,3 +226,205 @@ class UserChangePasswordView(APIView):
                 return Response(data={"state": "요청하신 데이터의 유저가 존재하지 않습니다"}, status=status.HTTP_404_NOT_FOUND)
         else:  # id 가 존재하지 않다면
             return Response(data={"state": "id 정보를 입력해주세요"}, status=status.HTTP_404_NOT_FOUND)
+
+
+# 게시글 관련 API
+class QuestionView(APIView):
+    def get(self, request, id):
+        # 질문에 대한 정보를 필터링 하기위한 파라미터
+        # 질문에 관한 데이터만 가져온다
+        question = question_board.objects.get(id=id)
+        # 질문에 대한 데이터 의 직렬화
+        question_serializer = QuestionDetailSerializer(question)
+
+        return Response(data=question_serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request):
+        serializer = QuestionCreateSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.create(validated_data=serializer.validated_data)
+            return Response(data={'state':"게시글이 정상적으로 작성되었습니다."},
+                            status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, id):
+        question = question_board.objects.get(id=id)  # user , title , body , state create
+        serializer = QuestionUpdateSerializer(question, data=request.data)
+
+        if serializer.is_valid():
+            serializer.update(question, serializer.validated_data)
+            return Response(data={"state":"게시글이 정상적으로 수정되었습니다"},status=status.HTTP_200_OK)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, id):
+        user = question_board.objects.get(id=id)
+        user.delete()
+        return Response(data={"state":"게시글이 정상적으로 삭제 되었습니다."},status=status.HTTP_204_NO_CONTENT)
+
+class QuestionListView(APIView):
+    def get(self, request , id):
+        queryset = question_board.objects.filter(user=id)
+        serializer = QuestionListSerializer(queryset, many=True)
+        return Response(data=serializer.data, status=status.HTTP_200_OK)
+
+
+
+##########################################################################
+################################ Paper ###################################
+##########################################################################
+class PaperTypeSituationView(APIView):
+    '''
+    GET: 타입(단어, 문장)이 들어오면 DB에 있는 상황(은행, 학교, 병원)들을 반환해준다.
+    '''
+
+    def get(self, request, type):
+        try:
+            qs = paper.objects.filter(type=type).distinct().values('situation')
+            serializer = PaperTypeSituationSerializer(qs,many=True)
+            return Response(serializer.data)
+        except paper.DoesNotExist:
+            return Response({"error": f"'{type}' 유형에 대한 데이터가 없습니다."}, status=status.HTTP_404_NOT_FOUND)
+
+
+class PaperTypeSituationChapterView(APIView):
+    '''
+    GET: 타입(단어, 문장)과 상황 들어오면 챕터 정보를 반환해준다.
+    '''
+
+    def get(self, request, type, situation):
+        try:
+            qs = paper.objects.filter(type=type, situation=situation).distinct().values('chapter')
+            serializer = PaperTypeSituationChapterSerializer(qs, many=True)
+            return Response(serializer.data)
+        except paper.DoesNotExist:
+            return Response({"error": f"'{type}' 유형과 '{situation}' 상황에 대한 데이터가 없습니다."},
+                            status=status.HTTP_404_NOT_FOUND)
+
+
+class PaperOneDataView(APIView):
+    '''
+    GET: 문제 번호가 들어오면 해당 문제를 반환한다.
+    '''
+
+    def get(self, request, id):
+        try:
+            qs = paper.objects.get(id=int(id))
+            serializer = PaperDataSerializer(qs)
+            return Response(serializer.data)
+        except paper.DoesNotExist:
+            return Response({"error": f"ID '{id}'에 대한 데이터가 없습니다."}, status=status.HTTP_404_NOT_FOUND)
+
+
+class PaperManyDataView(APIView):
+    '''
+    GET: 특정 유형(문장,단어,자음모음)와 상황(병원,학교)와 챕터가 오면 그 챕터에 속한 문제들을 반환한다.
+    '''
+
+    def get(self, request, type, situation, chapter):
+        try:
+            qs = paper.objects.filter(type=type, situation=situation, chapter=chapter)
+            serializer = PaperDataSerializer(qs, many=True)
+            return Response(serializer.data)
+        except paper.DoesNotExist:
+            return Response({"error": f"'{type}' 유형, '{situation}' 상황, '{chapter}' 챕터에 대한 데이터가 없습니다."},
+                            status=status.HTTP_404_NOT_FOUND)
+        except ValueError:
+            return Response({"error": "챕터는 정수값이어야 합니다."}, status=status.HTTP_400_BAD_REQUEST)
+
+
+##########################################################################
+############################ PracticeNote ################################
+##########################################################################
+class PracticeNoteView(APIView):
+    '''
+    GET: 특정 유저 + 특장 싱황 + 특정 챕터별 틀린문제를 return
+
+    POST: 새로운 문제를 푼 경우 DB에 정답(오답) 기록을 추가한다.
+
+    PUT: 기존 오답 이였던 문제를 정답으로 바꾼다.
+    '''
+
+    def get(self, request):
+        type = request.GET.get('type')
+        situation = request.GET.get('situation')
+        chapter = request.GET.get('chapter')
+        user_id = request.GET.get('user_id')
+
+
+        try:
+            practice_notes = practice_note.objects.select_related('paper').filter(
+                paper__type=type,
+                paper__situation=situation,
+                paper__chapter=chapter,
+                user=user_id,
+                is_answer=False
+            )
+            serializer = PracticeNoteSerializer(practice_notes, many=True)
+            return Response(serializer.data)
+        except practice_note.DoesNotExist:
+            return Response({"error": "해당하는 데이터가 없습니다."}, status=status.HTTP_404_NOT_FOUND)
+
+    def post(self, request):    # 중복 문제 발생 -> 한 문제 를 같은 사람이 여러번 푸는것이 record 에 남는다
+        serializer = PracticeNoteSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, paper_id, user_id):
+        try:
+            practice_note_instance = practice_note.objects.get(paper=paper_id, user=user_id)
+        except practice_note.DoesNotExist:
+            return Response({"error": f"paper_id '{paper_id}', user_id '{user_id}'에 대한 데이터가 없습니다."},
+                            status=status.HTTP_404_NOT_FOUND)
+
+        practice_note_instance.is_answer = True
+        practice_note_instance.save()
+
+        serializer = PracticeNoteSerializer(practice_note_instance)
+        return Response(serializer.data)
+
+
+##########################################################################
+############################ 문제 비율 확인 ################################
+##########################################################################
+class PaperStatesView(APIView):
+    '''
+    GET : 특정 유형,상황,챕터의 문제 개수와 정답 문제 개수를 전송합니다.
+    문제 개수와 정답 문제 개수가 같은 경우 틀린문제가 없으며 해당 챕터를 완강표시 합니다.
+    문제 개수와 정답 문제 개수가 다른 경우 풀지 않은 문제가 있거나 오답인 문제가 존재합니다.
+    '''
+
+    def get(self, request):
+        type = request.GET.get('type')
+        situation = request.GET.get('situation')
+        chapter = request.GET.get('chapter')
+        user_id = request.GET.get('user_id')
+
+        try:
+            practice_notes = practice_note.objects.select_related('paper').filter(
+                paper__type=type,
+                paper__situation=situation,
+                paper__chapter=chapter,
+                user=user_id,
+                is_answer=True
+            )
+            serializer = PracticeNoteSerializer(practice_notes, many=True)
+
+            answer_num = len(serializer.data)
+
+            qs = paper.objects.filter(
+                type=type,
+                situation=situation,
+                chapter=chapter
+            )
+            serializer = PaperDataSerializer(qs, many=True)
+            total_paper_num = len(serializer.data)
+
+            response_data = {
+                "answer_num": answer_num,
+                "total_paper_num": total_paper_num,
+            }
+            return Response(response_data)
+        except practice_note.DoesNotExist:
+            return Response({"error": "해당하는 데이터가 없습니다."}, status=status.HTTP_404_NOT_FOUND)
